Completed chapter 12.
Added UserRepository interface to handle database
methods for adding/deleting etc User objects in
a database.

Added UserService class to handle the business
logic of actually saving a User object to the
database.

Updated User class to be a Spring Entity for use
with a database.

Added UserService class to the UserController and
use it to save the User object to the database.

Added new test to cover saving a User to the
database. Added clean up step to clear the
database after each test.