Completed chapter 22-26
Added backend user validation and tests.

Added custom validator annotation and implementation to test for duplicate
usernames. A side effect of this change prevents the use of the validators
by hibernate prior to pushing data into the database.  This second layer of
validation has been disabled in the application config.